Modified Profile: [OrderBot] Il Mheg Side Quests.xml.
Attempt blind fix of quest: Scents of Security. Moved all interaction into the CodeChunk to stop TalkTo thinking the object doesn't exist when transitioning from not having used the item.

Added a wait before attempting the custom navigation for Quest: A Leafman and a Hero. Characters stuck in a conversation lock will fail the mount, or climb to the first vector prematurely missing the distance check.